Renamed directory FormElements to UIElements - Changed all the files imports to point to the new name of the directory - Moved InfoBox ton the UIElements directory
